requirement.txt
requests
matplotlib
json
re
os
jieba
wordcloud
Pillow
numpy
collections
sklearn
pandas
snownlp
beautifulsoup4